+++ packages/node/src/indexer/project.service.ts
@@ -6,6 +6,7 @@ import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
   PoiService,
   MmrService,
+  MmrQueryService,
   BaseProjectService,
   StoreService,
   NodeConfig,
@@ -36,6 +37,7 @@ export class ProjectService extends BaseProjectService<
     apiService: AlgorandApiService,
     poiService: PoiService,
     mmrService: MmrService,
+    mmrQueryService: MmrQueryService,
     sequelize: Sequelize,
     @Inject('ISubqueryProject') project: SubqueryProject,
     storeService: StoreService,
@@ -49,6 +51,7 @@ export class ProjectService extends BaseProjectService<
       apiService,
       poiService,
       mmrService,
+      mmrQueryService,
       sequelize,
       project,
       storeService,